+++ ReleaseNotes/release_notes.md
@@ -32,6 +32,7 @@
   1. further imporvements 
   1. make a try to fix morphing 
   1. fix for 3D-reweighting, add 3D-reweigtni test/example for  Paula Garcia 
+  1. add (self)addition/subttraction oprators for polyniias classes (`Polynomial`, `ChebyshevSum`, `LegendreSum`, `LegendreSum2`, `LegendreSum3` and `LegendreSum4`) with the same domain. 
 
 ## Backward incompatible: